HSEARCH-3614 Move ReflectionUtils and GenericTypeContext from the POJO mapper module to the utils module
They are general utils related to the Java language and not to POJO
specifically, and thus could theoretically be used in any module.
