+++ util/common/src/main/java/org/hibernate/search/util/common/reflect/impl/ReflectionUtils.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.util.impl;
+package org.hibernate.search.util.common.reflect.impl;
 
 import java.lang.reflect.Array;
 import java.lang.reflect.GenericArrayType;
