+++ util/common/src/test/java/org/hibernate/search/util/common/reflect/impl/CustomBoundedGenericInterface.java
@@ -4,8 +4,8 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.util.impl;
+package org.hibernate.search.util.common.reflect.impl;
 
 @SuppressWarnings("unused")
-interface CustomGenericInterface<T, R> {
+interface CustomBoundedGenericInterface<T extends Number & Cloneable> {
 }
