+++ util/common/src/test/java/org/hibernate/search/util/common/reflect/impl/CustomGenericType.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.util.impl;
+package org.hibernate.search.util.common.reflect.impl;
 
 // Switch the type parameters on purpose
 final class CustomGenericType<T, R> implements CustomGenericInterface<R, T> {
