+++ util/common/src/test/java/org/hibernate/search/util/common/reflect/impl/GenericTypeContextTest.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.util.impl;
+package org.hibernate.search.util.common.reflect.impl;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.fail;
@@ -17,9 +17,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import org.hibernate.search.mapper.pojo.testsupport.util.TypeCapture;
-import org.hibernate.search.mapper.pojo.testsupport.util.WildcardTypeCapture;
-import org.hibernate.search.mapper.pojo.testsupport.util.WildcardTypeCapture.Of;
+import org.hibernate.search.util.impl.test.reflect.TypeCapture;
+import org.hibernate.search.util.impl.test.reflect.WildcardTypeCapture;
+import org.hibernate.search.util.impl.test.reflect.WildcardTypeCapture.Of;
 
 import org.junit.Rule;
 import org.junit.Test;
